Rename program_supervisor, load program listing from JSON config (#176)
