+++ modules/core/protos/CMakeLists.txt
@@ -2,4 +2,5 @@ farm_ng_add_protobufs(farm_ng_core_protobuf
   PROTO_FILES
   ./farm_ng/core/io.proto
   ./farm_ng/core/resource.proto
+  ./farm_ng/core/programd.proto
   )
