+++ modules/core/protos/farm_ng/core/programd.proto
@@ -2,13 +2,27 @@ syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
 
-package farm_ng.frontend;
-option go_package = "github.com/farm-ng/genproto/frontend";
+package farm_ng.core;
+option go_package = "github.com/farm-ng/genproto/core";
+
+message ProgramdConfig {
+  repeated Program programs = 1;
+}
 
 message Program {
   string id = 1;
   string name = 2;
   string description = 3;
+
+  message LaunchPath {
+    string path = 1;
+    // If set, the path will be resolved relative to the
+    // value of this environment variable.
+    string root_env_var = 2;
+  }
+
+  LaunchPath launch_path = 4;
+  repeated string launch_args = 5;
 }
 
 message ProgramExecution {
