+++ modules/frontend/frontend/src/registry/events.ts
@@ -42,7 +42,7 @@ import {
   ProgramSupervisorStatus,
   StartProgramRequest,
   StopProgramRequest,
-} from "@farm-ng/genproto-frontend/farm_ng/frontend/program_supervisor";
+} from "@farm-ng/genproto-core/farm_ng/core/programd";
 import {
   CalibrateIntrinsicsConfiguration,
   CalibrateIntrinsicsResult,
@@ -169,13 +169,15 @@ export const eventRegistry = inferKeys({
   "type.googleapis.com/farm_ng.core.Event": BusEvent,
   "type.googleapis.com/farm_ng.core.LoggingCommand": LoggingCommand,
   "type.googleapis.com/farm_ng.core.LoggingStatus": LoggingStatus,
-  "type.googleapis.com/farm_ng.frontend.ProgramSupervisorStatus": ProgramSupervisorStatus,
-  "type.googleapis.com/farm_ng.frontend.StartProgramRequest": StartProgramRequest,
-  "type.googleapis.com/farm_ng.frontend.StopProgramRequest": StopProgramRequest,
+  "type.googleapis.com/farm_ng.core.ProgramSupervisorStatus": ProgramSupervisorStatus,
+  "type.googleapis.com/farm_ng.core.StartProgramRequest": StartProgramRequest,
+  "type.googleapis.com/farm_ng.core.StopProgramRequest": StopProgramRequest,
   "type.googleapis.com/farm_ng.perception.ApriltagConfig": ApriltagConfig,
   "type.googleapis.com/farm_ng.perception.ApriltagDetections": ApriltagDetections,
   "type.googleapis.com/farm_ng.perception.ApriltagRig": ApriltagRig,
   "type.googleapis.com/farm_ng.perception.CameraConfig": CameraConfig,
+  "type.googleapis.com/farm_ng.perception.CameraPipelineCommand": CameraPipelineCommand,
+  "type.googleapis.com/farm_ng.perception.CameraPipelineConfig": CameraPipelineConfig,
   "type.googleapis.com/farm_ng.perception.CaptureVideoDatasetConfiguration": CaptureVideoDatasetConfiguration,
   "type.googleapis.com/farm_ng.perception.CaptureVideoDatasetResult": CaptureVideoDatasetResult,
   "type.googleapis.com/farm_ng.perception.CaptureVideoDatasetStatus": CaptureVideoDatasetStatus,
@@ -188,8 +190,6 @@ export const eventRegistry = inferKeys({
   "type.googleapis.com/farm_ng.perception.NamedSE3Pose": NamedSE3Pose,
   "type.googleapis.com/farm_ng.perception.SE3Pose": SE3Pose,
   "type.googleapis.com/farm_ng.perception.TagConfig": TagConfig,
-  "type.googleapis.com/farm_ng.perception.CameraPipelineCommand": CameraPipelineCommand,
-  "type.googleapis.com/farm_ng.perception.CameraPipelineConfig": CameraPipelineConfig,
   "type.googleapis.com/farm_ng.perception.Vec2": Vec2,
   "type.googleapis.com/farm_ng.perception.VideoFileCamera": VideoFileCamera,
   "type.googleapis.com/farm_ng.tractor.MotorControllerState": MotorControllerState,
