+++ modules/frontend/protos/CMakeLists.txt
@@ -1,5 +1,4 @@
 farm_ng_add_protobufs(farm_ng_frontend_protobuf
   PROTO_FILES
-  ./farm_ng/frontend/program_supervisor.proto
   ./farm_ng/frontend/webrtc_proxy_service.proto
   )
